microservices/token_info/requirements.txt
pandas==2.0.0
python-dotenv
Requests
SQLAlchemy
pymysql
